refactor: 입력과 입력 문자를 게임 숫자 변환하는 기능 분리
