+++ src/main/java/baseball/console/GameConsole.java
@@ -1,0 +1,10 @@
+package baseball.console;
+
+import camp.nextstep.edu.missionutils.Console;
+
+public class GameConsole {
+
+    public String input() {
+        return Console.readLine();
+    }
+}
